fix-report-values > Hardcore refactor processing
\### Rationale
Refactored data extract functionality to separate it from creation of
a GitLab security report. This ensures code reuse and better
maintainability for when GitLab schema updates. Developers now only
have to go to the utils section to add a new field in relation to the
report and find the data necessary.  It required too much work &
thought of the developer to maintain matches across basically 2 report
parsers.

The abstraction here now keeps the same template across report types
and only worry about how to generate the values needed while the util
functions generate the GitLab expected schema format.  This also
helped centralize the hard coded NPM values.

In the process of this commit, the report v1 parser was upgraded to
extract remediations and pass them along to GitLab.  This was likely
overlooked since Gitlab's format did not include remediations
initially. This project is now up-to-date with v14.0.3 of the
dependency-scanning report schema definition.

Added a bunch of comments to document the new construct.