fix-report-values > v1-snap: Solve schema errors
\### Rationale
Now that the unit-testing validates the schemas against the GitLab
schema object via `jsonschema`'s validator. There were multiple
conflicts found that the reports were invalid that were being
generated.

In line with the new functionality abstraction, the fields are now
sorted the same as v2 reports and also are similar in content as well