pre-release > Bump `jest` & `babel-jest` to `v27`
\### Rationale
Update dependencies for more functionality and better workflows.
`Jest@v27` included `jest-circus` which resolves some of the deep flaws
in the original `jasmine` when it comes to error handling with the
`before*()` hooks which would still run the dependent testcase even if
the setup functionality errored. Was able to drop the additional manual
check I implemented since now it has been proven to not be an issue
any longer. `Jest@v27` will stop any conditional tests if a before hook
fails and the afterhook will still run.